% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulemodels.R
\name{cbaCSV}
\alias{cbaCSV}
\title{Example workflow that reads a data file from csv, learns a cba rule set and saves the resulting rule set back to csv.}
\usage{
cbaCSV(path, outpath = NULL, classatt = NULL, idcolumn = NULL,
  rulelearning_options = NULL, pruning_options = NULL)
}
\arguments{
\item{path}{path to csv file with data.}

\item{outpath}{path to write the rule set to.}

\item{classatt}{the name of the class attribute.}

\item{idcolumn}{the name of the id column in the dataf ile.}

\item{rulelearning_options}{custom options for the rule learning algorithm overriding the default values.}

\item{pruning_options}{custom options for the pruning algorithm overriding the default values.}

\item{target_rule_count}{the number of initial rules before pruning.}
}
\value{
List of items. For example, for input column \code{Species} with distinct values \code{versicolor, setosa} returns a list of two items  \code{Species=versicolor, Species=setosa}
}
\description{
Example workflow that reads a data file from csv, learns a cba rule set and saves the resulting rule set back to csv.
}
\examples{
\\code{
  cbaCSV("data/KO_Bank_all.csv")

}


}

