% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulemodels.R
\name{cbaIris}
\alias{cbaIris}
\title{Test cba workflow on iris dataset.}
\usage{
cbaIris()
}
\value{
Accuracy.
}
\description{
Test workflow on iris dataset: learns a cba classifier on one "train set" part , and applies it to the second  "test set" part.
}

