% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{CBARuleModelAccuracy}
\alias{CBARuleModelAccuracy}
\title{Prediction Accuracy}
\usage{
CBARuleModelAccuracy(prediction, groundtruth)
}
\arguments{
\item{prediction}{a vector with predictions}

\item{groundtruth}{a vector with groundtruth}
}
\value{
Accuracy
}
\description{
Compares predictions with groundtruth and outputs accuracy.
}
