% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulemodels.R
\name{default_pruning_setup}
\alias{default_pruning_setup}
\title{Default parameters for pruning}
\usage{
default_pruning_setup()
}
\value{
A list with default parameter valuees
}
\description{
Default parameters for pruning
}

