% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulemodels.R
\name{default_rule_learning_setup}
\alias{default_rule_learning_setup}
\title{Default parameters for rule learning.}
\usage{
default_rule_learning_setup()
}
\value{
A list with default parameter valuees
}
\description{
Default parameters for rule learning.
}

