% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlp2.R
\name{discretizeUnsupervised}
\alias{discretizeUnsupervised}
\title{Unsupervised Discretization}
\usage{
discretizeUnsupervised(
  data,
  labels = FALSE,
  infinite_bounds = FALSE,
  categories = 3,
  method = "cluster"
)
}
\arguments{
\item{data}{input numeric vector.}

\item{labels}{a logical indicating whether the bins of the discretized data should be represented by integer codes or as interval notation using (a;b] when set to TRUE.}

\item{infinite_bounds}{a logical indicating how the bounds on the extremes should look like.}

\item{categories}{number of categories (bins) to produce.}

\item{method}{clustering method, one of "interval" (equal interval width), "frequency" (equal frequency), "cluster" (k-means clustering). See also documentation of the \code{\link[arules]{discretize}} function from the arules package.}
}
\value{
Discretized data. If there was no discretization specified for some columns, these are returned as is.
}
\description{
Discretizes provided numeric vector.
}
\examples{
  discretizeUnsupervised(datasets::iris[[1]])

}
