% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlp2.R
\name{discrNumeric}
\alias{discrNumeric}
\title{Method that matches rule model against test data.}
\usage{
discrNumeric(df, classatt, min_distinct_values = 3, unsupervised_bins = 3,
  discretize_class = FALSE)
}
\arguments{
\item{df}{a data frame with data.}

\item{classatt}{name the class attribute in df}

\item{min_distinct_values}{the minimum number of unique values a column needs to have to be subject to supervised discretization.}

\item{unsupervised_bins}{number of target bins for  discretizing the class attribute. Ignored when the class attribute is not numeric or when discretize_class is set to FALSE.}

\item{discretize_class}{logical value indicating whether the class attribute should be discretized. Ignored when the class attribute is not numeric.}
}
\value{
list with two slots: $cutp with cutpoints and $Disc.data with discretization results
}
\description{
Method that matches rule model against test data.
}
\examples{
  #data(iris)
  discrNumeric(iris,"Species")

}

