% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{getAppearance}
\alias{getAppearance}
\title{Method that generates items for values in given data frame column.}
\usage{
getAppearance(df, classatt)
}
\arguments{
\item{df}{a data frame contain column \code{classatt}.}

\item{classatt}{name of the column in \code{df} to generate items for.}
}
\value{
appearance object for mining classification rules
}
\description{
Method that generates items for values in given data frame column.
}
\examples{
utils::data(iris)
getAppearance(iris,"Species")

}

