% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulemodels.R
\name{getItems}
\alias{getItems}
\title{Method that generates items for values in given data frame column.}
\usage{
getItems(df, classatt)
}
\arguments{
\item{df}{a data frame contain column \code{classatt}.}

\item{classatt}{name of the column in df to generate items for.}
}
\value{
a list of items.
}
\description{
Method that generates items for values in given data frame column.
}
\examples{
#For input column \\code{Species} with distinct values \\code{versicolor, setosa},
#the function returns a list of two items  \\code{Species=versicolor, Species=setosa}.
#data(iris)
  getItems(iris,"Species")


}

