% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{humtemp}
\alias{humtemp}
\title{Comfort level based on temperature and humidity of the environment}
\format{A data frame with 34 rows and 3 variables}
\usage{
data(humtemp)
}
\description{
A syntetic toy dataset. The variables are as follows:
}
\details{
\itemize{
  \item Temperature.
  \item Humidity.
  \item Class. Comfort level
}
}
\keyword{humtemp}
