% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulemodels.R
\docType{methods}
\name{ruleMatch,RuleModel,data.frame-method}
\alias{ruleMatch,RuleModel,data.frame-method}
\title{Title}
\usage{
\S4method{ruleMatch}{RuleModel,data.frame}(rule_model, test)
}
\arguments{
\item{rule_model}{RuleModel.}

\item{test}{data.frame.}
}
\value{
A vector with predictions.
}
\description{
Title
}
\seealso{
\link{cbaIris}
}

