% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulemodels.R
\name{ruleMatch}
\alias{ruleMatch}
\title{method that matches rule model against test data}
\usage{
ruleMatch(rule_model, test)
}
\arguments{
\item{rule_model}{a \link{RuleModel} class instance}

\item{test}{a data frame with test data, the data frame may optionally contain the target class attribute - if it does, the method also computes and <i>prints</i> accuracy.}
}
\value{
factor with predictions for input instances
}
\description{
method that matches rule model against test data
}
\examples{
\\code{
  data(iris)
  train<-iris[1:100,]
  test<-iris[101:length(iris),]
  increase for more accurate results in longer time
  target_rule_count<-1000
  rm<-cba(test<-iris[101:length(iris),],target_rule_count=1000,classatt="Species")
  ruleMatch(rm,test)
}
}
\seealso{
Example \code{\link{learnprune_iris}}
}

