% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\docType{class}
\name{RuleModel-class}
\alias{RuleModel}
\alias{RuleModel-class}
\title{RuleModel}
\description{
This class represents a rule-based classifier.
}
\section{Slots}{

\describe{
\item{\code{rules}}{an object of class rules from arules package}

\item{\code{cutp}}{list of cutpoints}

\item{\code{classatt}}{name of the target class attribute}
}}

