% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulemodels.R
\name{cba}
\alias{cba}
\title{example workflow that reads a data file from csv, learns a cba rule set and saves the resulting rule set back to csv.}
\usage{
cba(train, classatt, rulelearning_options = NULL, pruning_options = NULL)
}
\arguments{
\item{train}{a data frame with data.}

\item{classatt}{the name of the class attribute.}

\item{rulelearning_options}{custom options for the rule learning algorithm overriding the default values.}

\item{pruning_options}{custom options for the pruning algorithm overriding the default values.}
}
\value{
Object of class \link{RuleModel}.
}
\description{
example workflow that reads a data file from csv, learns a cba rule set and saves the resulting rule set back to csv.
}
\examples{
  cba(iris,"Species",rulelearning_options=list(target_rule_count=50000))
  data(iris)
}

