% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{cbaCSV}
\alias{cbaCSV}
\title{Example CBA Workflow with CSV Input}
\usage{
cbaCSV(
  path,
  outpath = NULL,
  classAtt = NULL,
  idcolumn = NULL,
  rulelearning_options = NULL,
  pruning_options = NULL
)
}
\arguments{
\item{path}{path to csv file with data.}

\item{outpath}{path to write the rule set to.}

\item{classAtt}{the name of the class attribute.}

\item{idcolumn}{the name of the id column in the dataf ile.}

\item{rulelearning_options}{custom options for the rule learning algorithm overriding the default values.}

\item{pruning_options}{custom options for the pruning algorithm overriding the default values.}
}
\value{
Object of class \link{CBARuleModel}
}
\description{
Learns a CBA rule set and saves the resulting rule set back to csv.
}
\examples{
 # cbaCSV("path-to-.csv")


}
