% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{cbaIris}
\alias{cbaIris}
\title{Test CBA Workflow on Iris Dataset}
\usage{
cbaIris()
}
\value{
Accuracy.
}
\description{
Test workflow on iris dataset: learns a cba classifier on one "train set" fold , and applies it to the second  "test set" fold.
}
