% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cba.R
\name{CBARuleModelAccuracy}
\alias{CBARuleModelAccuracy}
\title{Prediction Accuracy}
\usage{
CBARuleModelAccuracy(prediction, groundtruth)
}
\arguments{
\item{prediction}{vector with predictions}

\item{groundtruth}{vector with true labels}
}
\value{
Accuracy
}
\description{
Compares predictions with true labels and outputs accuracy.
}
