% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulemodels.R
\name{getItems}
\alias{getItems}
\title{Method that generates items for values in given data frame column.}
\usage{
getItems(df, classatt)
}
\arguments{
\item{df}{a data frame contain column \link{classatt}.}

\item{classatt}{name of the column in df to generate items for.}
}
\value{
a list of items.
}
\description{
Method that generates items for values in given data frame column.
}
\examples{
For input column \\texttt{Species} with distinct values \\texttt{versicolor, setosa}, the function returns a list of two items  \\texttt{Species=versicolor, Species=setosa}.
\\code{
  data(iris)
  getItems(iris,"Species")
}


}

