% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulemodels.R
\docType{class}
\name{RuleModel}
\alias{RuleModel}
\title{RuleModel}
\description{
This class represents a rule-based classifier.
}
\details{
\describe{
   \item{rules}{An object of class rules}

\item{cutp}{A list of cutpoints defining required discretization for the rule set to be applicable}

\item{classatt}{The name of the class attribute}
 }
}
\section{Slots}{

\describe{
\item{\code{rules}}{an object of class rules from arules package}

\item{\code{cutp}}{list of cutpoints}

\item{\code{classatt}}{name of the target class attribute}
}}

